+++ src/ASNeG-Demo/Library/CreateDeleteNode.h
@@ -1,5 +1,5 @@
 /*
-   Copyright 2017-2018 Kai Huebl (kai@huebl-sgh.de)
+   Copyright 2018 Kai Huebl (kai@huebl-sgh.de)
 
    Lizenziert gemäß Apache Licence Version 2.0 (die „Lizenz“); Nutzung dieser
    Datei nur in Übereinstimmung mit der Lizenz erlaubt.
@@ -15,16 +15,12 @@
    Autor: Kai Huebl (kai@huebl-sgh.de)
  */
 
-#ifndef __OpcUaServerApplicationDemo_Event_h__
-#define __OpcUaServerApplicationDemo_Event_h__
+#ifndef __OpcUaServerApplicationDemo_CreateDeleteNode_h__
+#define __OpcUaServerApplicationDemo_CreateDeleteNode_h__
 
 #include "OpcUaStackCore/Utility/IOThread.h"
-#include "OpcUaStackCore/Application/ApplicationMethodContext.h"
-#include "OpcUaStackCore/Application/ApplicationEventItemStartContext.h"
-#include "OpcUaStackCore/Application/ApplicationEventItemStopContext.h"
 #include "OpcUaStackServer/Application/ApplicationIf.h"
 #include "OpcUaStackServer/Application/ApplicationInfo.h"
-#include "OpcUaStackServer/AddressSpaceModel/BaseNodeClass.h"
 
 using namespace OpcUaStackCore;
 using namespace OpcUaStackServer;
@@ -32,35 +28,22 @@ using namespace OpcUaStackServer;
 namespace OpcUaServerApplicationDemo
 {
 
-	class Event
+	class CreateDeleteNode
 	{
 	  public:
-		Event(void);
-		~Event(void);
+		CreateDeleteNode(void);
+		~CreateDeleteNode(void);
 
 		bool startup(IOThread& ioThread, ApplicationServiceIf& applicationServiceIf, ApplicationInfo* applicationInfo);
 		bool shutdown(void);
-
-	  private:
-		void eventItemStartCallback(ApplicationEventItemStartContext* applicationEventItemStartContext);
-		void eventItemStopCallback(ApplicationEventItemStopContext* applicationEventItemStopContext);
-
-		bool registerEventCallbacks(void);
-		bool getNamespaceInfo(void);
-		void startTimerLoop(void);
 		void timerLoop(void);
 
-		void sendEvent11(void);
-		void sendEvent12(void);
-		void sendEvent21(void);
-
+	  private:
 		IOThread* ioThread_;
-		SlotTimerElement::SPtr slotTimerElement_;
 		ApplicationServiceIf* applicationServiceIf_;
 		ApplicationInfo* applicationInfo_;
-
-		uint32_t namespaceIndex_;
-		uint32_t counter_;
+		SlotTimerElement::SPtr slotTimerElement_;
+		bool nodesExist_;
 	};
 
 }