+++ src/ASNeG-Demo/Library/DemoLibrary.cpp
@@ -37,6 +37,7 @@ namespace OpcUaServerApplicationDemo
 	, authentication_()
 	, historicalAccess_()
 	, discovery_()
+	, createDeleteNode_()
 	{
 	}
 
@@ -60,6 +61,7 @@ namespace OpcUaServerApplicationDemo
 		authentication_.startup(ioThread_, service(), applicationInfo());
 		historicalAccess_.startup(ioThread_, service(), applicationInfo());
 		discovery_.startup(ioThread_, service(), applicationInfo());
+		createDeleteNode_.startup(ioThread_, service(), applicationInfo());
 		return true;
 	}
 
@@ -68,6 +70,7 @@ namespace OpcUaServerApplicationDemo
 	{
 		Log(Debug, "DemoLibrary::shutdown");
 
+		createDeleteNode_.shutdown();
 		discovery_.shutdown();
 		historicalAccess_.shutdown();
 		authentication_.shutdown();