+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Release 3.3.0
 
 **Features**:
 
+* refactoring application service interface #7, by @huebl
+
 **Bugs**:
  
 **Documentation**: